TLtest
this is the code that i've been working on. it has the a UNet structure incorporating VGG16 weights for transfer learning. if you are able to run the code,  the model_TL_UNET.h5 file that is used for prediction function is saved in images/mask. i couldn't the model_TL_UNET.h5 in this commit. i'm not sure why.
Please message me for any questions. This is trial environment set-up for our actual dataset.